combine test_csv_io and test_cui to test_ui
