+++ tests/test_ui.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Created on Fri Feb  1 20:48:55 2019
+Created on Sun Mar 10 21:46:07 2019
 
 @author: LiNaK
 """
@@ -9,7 +9,6 @@ Created on Fri Feb  1 20:48:55 2019
 import unittest
 from unittest.mock import patch
 from pathlib import Path
-import sys
 import os
 
 # 3rd party's module
@@ -25,13 +24,35 @@ from testing_utility.unittest_util import cls_startstop_msg as add_msg
 from test_utility import (Inheration_test_base)
 
 # target class
+from src.dataio.cui import ValueAskCUI
 import src.dataio.csvio as csvio
-from src.dataio.csvio import CSVIO
 from src.dataio.io_com import (PathDialog)
 
+
 # interface
+from src.interface.intfc_com import (ValueAsk)
 from src.interface.intfc_com import (Reader, Writer, PathAsk)
 
+@add_msg
+class TestCSVIOInterfaces(Inheration_test_base,unittest.TestCase):
+    # Test inheration of interfaces.
+    _sub_sup_class_pairs =((ValueAskCUI, ValueAsk),
+                           )
+@add_msg
+class Test_CUI_ValueAsk(unittest.TestCase):
+    def test_data_reading(self):
+        self.vac = ValueAskCUI()
+        if __name__=='__main__':
+            '''
+            This is only active by main.
+            This is irregular test because user interface. Then check by hand.
+            '''
+            
+            inputval = 'input'
+            msg = 'Please input "{}" to test {} class.'\
+            .format(inputval, self.vac.__class__)
+            val = self.vac.get_value(msg)
+            self.assertEqual(inputval, val)
 
 @add_msg
 class TestCSVIOInterfaces(Inheration_test_base,unittest.TestCase):
@@ -90,7 +111,7 @@ class DummyDataForTest():
         dummy_data_foleder_name = 'dummy_data'
         _dummy_data_file_name = 'test_csv_writer.csv'
         return str(module_path / dummy_data_foleder_name / _dummy_data_file_name)
-
+        
 
 if __name__=='__main__':
     unittest.main()
